requirements.txt
openai>=1.0.0
anthropic>=0.5.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
pydantic>=2.0.0
fastapi>=0.100.0
uvicorn>=0.22.0
pytest>=7.4.0
python-multipart>=0.0.6 